Updated the Logo and No More Ads.
Had this thing on the website about Ads, removed it. I'm not here to profit lmfao.
Secondly, I had this "Logo" thing that came with the Bootstrap ting. So I modified it to say "Nathan's Techtimeline".
Pressing on it will bring you back to index.html (Which is the main website)
that's all <3